copy fields, dynamic fields, defaults
This commit reflects some learning about copy fields, dynamic
fields and the fields that Solr searches by default.

Most of the learning is captured in updates to README.md

As regards dynamic fields, it is to be noted that tei.xsl
generates 'author_s' fields, despite the fact that the
schema definition for medieval does not include this
field. However documents with 'author_s' elements can be
indexed because the schema includes a dynamic field definition
with the pattern `*_s`. This enables the dynamic creation of
a field defintion for `author_s`.